Change whole project structure and way of generating floors
Previous structure was confusing, so I divided abstract classes from the rest. 
The way of generating floors was too buggy because yielding in this Godot version is a little bugged. I changed it completely, so it no longer uses json, but is entirely Node-based. Also room collision detection was entirely not working, so I fixed it.